third_party/blink/web_tests/external/wpt/css/css-inline/parsing/alignment-baseline-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property alignment-baseline value 'text-bottom'
  assert_true: 'text-bottom' is a supported value for alignment-baseline. expected true got false
[FAIL] Property alignment-baseline value 'text-top'
  assert_true: 'text-top' is a supported value for alignment-baseline. expected true got false
[FAIL] Property alignment-baseline value 'bottom'
  assert_true: 'bottom' is a supported value for alignment-baseline. expected true got false
[FAIL] Property alignment-baseline value 'center'
  assert_true: 'center' is a supported value for alignment-baseline. expected true got false
[FAIL] Property alignment-baseline value 'top'
  assert_true: 'top' is a supported value for alignment-baseline. expected true got false
Harness: the test ran to completion.
